c.2: hashmaps, vectors, hide MAL struct, reduce duplications, fix 62d1ed3c
Makefile:
  Compute the build flags with pkg-config.
  Make more settings available on the command line.
  Enable the -fanalyzer warning (and close the issue found in slurp).
core:
  Simplify the argument parsing and return value constructor in ".".
env:
  Implement with hashmaps. Allow printing by DEBUG-EVAL.
error (new):
  Format most error messages consistently.
  When an error is detected, show all arguments instead of only the
  remaining ones (this was not informative when the rest was empty).
  Restore the previous error prefix (the MAL function, not the C function).
  Use a global variables instead of a type, simplifying functions
  returning non-MAL types.
hashmap (rewritten, moved to ./):
  Replace binary trees with hashed maps.
linked_list (moved to ./):
  Return a signed count in order to spare a later conversion.
printer:
  Allow a width specifier (for DEBUG-EVAL).
  Print MAL type sets (for bad_type) and hashmaps (for DEBUG-EVAL).
  Shorten name for functions (else REPL fills the screen).
  Move commands outside of assert (they didn’t survive NDEBUG).
  Inverse the meaning of " " so that list *are* spaced by default.
readline (new):
  Avoid duplication in steps and core (core was incorrect).
steps:
  Properly report reader errors (this was dealt with by the null case in EVAL,
  but not documented and removed by 62d1ed3c).
  Stop preventing TCO when if returns nil (the benefit is dubious).
  Add a quasiquote_folder used by lists and vectors.
  Stop calling quasiquote twice on each list/vector element (error in 62d1ed3c).
types:
  Hide the struct into the .c file.
  Compute the hash of each string at creation.
  Allocate common symbols once at startup.
  Remove metadata from most types.
  Use the same "pattern-matching" function to test the type and extract the contents.
  Add metadata as an optional argument to some constructor.
vectors (new):
  Provide an iterator common with lists.

global
  Move equality from core to types, use it in steps and hashmaps.

  In closures, store the raw fn* arguments (The &more syntax without
  space is lost but could be restored if this seems important).

+++ impls/c.2/linked_list.c
@@ -11,9 +11,9 @@ list list_push(list lst, MalType data_ptr) {
   return new_head;
 }
 
-size_t list_count(list lst) {
+int list_count(list lst) {
 
-  size_t counter = 0;
+  int counter = 0;
 
   while(lst) {
